PGY V1.0.1 build1  Release -160806.0215pm

+++ MSPopGesture/Main/Me/Controllers/MSMeVC.m
@@ -181,7 +181,7 @@
 }
 - (void)clickedBubbleButton:(UIButton *)button
 {
-    NSLog(@"click:%ld", button.tag);
+    NSLog(@"click:%ld", (long)button.tag);
 }
